Move DirectSound plugin to its own folder
Will eventually need to separate the APIs and its classes into a generic class.
