+++ src/core/HLE/DSound/DirectSound/DirectSound.hpp
