+++ src/core/HLE/DSound/DirectSound/DirectSoundInline.hpp
